On sale tickets can now be put on sale
There was a problem with the 'put on sale' button in performances#index because it would put al tickets on sale and if any tickets were already on sale, the user would see ajax errors.  Now, on_sale to on_sale is a valid transition (obviously benign).  Changed the grouped form and ticket.on_salable? to reflect this logic.