Changed Comp to accept an array of ticket ids or an array of tickets
The tickets will be loaded form disk if ids are passed.  If Tickets are passed, the tickets will not be reloaded.  This change was made to acommodate box office, which has the actual ticket objects to be comped.  If Comp reloads those tickets, box office's copies are stale and the comp appears ot have failed.  All this is backward compatible with the current comp flow