Refactoring comps
Moved a lot of the validation logic into comp.rb to take the load off of the controller.  When you init comps now, you must include the benefactor.  Made changes to sale and comp flow.
Removed find_person because the flow was consolidated so that it goes new->confirm->create