More fumbling with the asset path
Since jammit jams javascript to assets, there is no helper to generate this link properly that would work with the escaped code in the widget genrator.  I tried raw and html_safe. So I rolled a asset_path helper to generate a javascript_path then gsub javascript with assets

This WILL NOT work locally and SHOULD NOT be used outside of the widget generator