Add partial-text searching.
I had to include the sold configuration folder, but
I'm still ignoring the data and pids folders. I also
created a spec/features folder to store the searching
spec. It needs to be in its own file, in order to
turn on web access and search indexing, necessary to
run the specs.