(OSE) UUID
Added uuid field to show and event.  This was necessary because the api feed smashes together reseller shows and shows.  Since those are different objects, they could conceivably have the same "id".  One implementor said that this was causing problems so I added a uuid for shows and events and put that into the feed.  Also added a multi_find method on both so that one can pass either the id or the uuid to find an object