(OSE) Item changes
Changed the refundable?, exchangeable? and returnable? criteria to look at the actual settlement_id instead of stat=settled because cash and fa donations are also settled.  We only want to prevent refunding cc transactions that we have paid the producer for

Also changed to_exchange! to inherit the state and sold_price of the item being exchanged