(OSE) Refactor OhNoes.
I turned it into a Rails Concern, which handles
mixins with a little more grace. Concerns also
handle dependencies a little better. They're not an
issue with this module, it's just nice to know.

I'm also using update_column here. I tested this
against all of the classes that use this module
just in case, and it's working while removing the
deprecated method.