(OSE) Add cart accessor to discount.
We need access to the cart throughout discounts.
Instead of continually passing around a cart
variable, we just store access to it locally. This
removes a lot of unnecessary code and makes it
easier to read.

In the future, we can possibly pass some of the
functionality off to the cart itself, but right now
this approach reduces coupling and simplifies the
code.